Refactor to make client and util common for different packages
