+++ src/main/scala/com/codacy/client/bitbucket/v1/service/PullRequestServices.scala
@@ -1,7 +1,7 @@
 package com.codacy.client.bitbucket.v1.service
 
-import com.codacy.client.bitbucket.v1.client.{BitbucketClient, Request, RequestResponse}
-import com.codacy.client.bitbucket.v1.util.CommitHelper
+import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.util.CommitHelper
 import com.codacy.client.bitbucket.v1.{PullRequest, PullRequestComment, SimpleCommit, SimplePullRequestComment, PullRequestReviewers}
 import play.api.libs.json._
