+++ src/main/scala/com/codacy/client/bitbucket/v1/service/RepositoryServices.scala
@@ -1,6 +1,6 @@
 package com.codacy.client.bitbucket.v1.service
 
-import com.codacy.client.bitbucket.v1.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
 import com.codacy.client.bitbucket.v1.{Repository, SimpleRepository, SshKey}
 import play.api.libs.json.Json
